+++ src/server/utils/selector.c
@@ -319,7 +319,7 @@ void selector_destroy(fd_selector s) {
         if (s->fds != NULL) {
             for (size_t i = 0; i < s->fd_size; i++) {
                 if (ITEM_USED(s->fds + i)) {
-                    selector_unregister_fd(s, i);
+                    selector_unregister_fd(s, (int)i);
                 }
             }
             pthread_mutex_destroy(&s->resolution_mutex);
